:recycle: Separate base egg animation from state
