+++ src/friendo/state/idle/egg.js
@@ -4,7 +4,7 @@
  */
 
 import State from '../state'
-import { STATS } from '../../constants'
+import AEgg from '../../animation/egg'
 
 export const ID = 'state_baby'
 
@@ -13,40 +13,9 @@ export default class Egg extends State {
     if (!savedState) savedState = {}
     super(savedState)
     this.id = ID
-    this.shaking = savedState.shaking || false
-
-    // Left blank to avoid speaking while and egg
-    this.phrasebook = () => ['']
-    this.words = ''
-
     this.returnTo = this.id
-  }
 
-  draw(g, x, y, friendo) {
-    super.draw(g, x, y, friendo)
-
-    // Shake if not currently shaking and with a probability
-    // based on egg level
-    const rand = Math.floor(Math.random() * 100)
-    const shake =
-      (friendo.getStat(STATS.EGG) < 2 ? 0 : ((friendo.getStat(STATS.EGG) - 1) * 10 >= rand))
-    if (!this.shaking && shake) {
-      this.shaking = true
-    }
-
-    // do a shake
-    if (this.shaking) {
-      this.shaking = false
-      return this.frame2(g, x, y, friendo)
-    }
-    return this.frame1(g, x, y, friendo)
-  }
-
-  frame1(g, x, y, friendo) {
-    friendo.element.drawEgg(g, x, y, friendo)
-  }
-
-  frame2(g, x, y, friendo) {
-    friendo.element.drawEgg(g, x + 2, y, friendo)
+    // phrasebook left blank to avoid speaking while and egg
+    this.anim = new AEgg(savedState.anim, () => [''])
   }
 }
